# internal/gateway/apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: Waveyard API
  version: 1.0.0
  description: |
    HTTP API for the Waveyard beat marketplace.

    Use **Authorize** to provide an access token as `Bearer <token>` for protected
    operations. Refresh tokens are stored in an HTTP-only `refresh_token` cookie.
  contact:
    name: Waveyard Engineering
servers:
  - url: /
    description: Current API host
security: []
...
paths:
...
  /me/favorites:
    get:
      tags: [Catalog]
      operationId: listMyFavorites
      summary: List the authenticated user's favorited beats
      description: |
        Returns favorited beats ordered by favorite creation time, newest first.
        The cursor is opaque and must be returned unchanged by clients.
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          description: Maximum number of favorites to return.
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20

        - name: cursor
          in: query
          description: Opaque cursor returned by the previous response.
          schema:
            type: string

        - $ref: "#/components/parameters/CloudflareCountry"
        - $ref: "#/components/parameters/CountryCode"

      responses:
        "200":
          description: Favorited beats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoritePage"

        "400":
          $ref: "#/components/responses/BadRequest"

        "401":
          $ref: "#/components/responses/Unauthorized"

        "500":
          $ref: "#/components/responses/InternalError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token returned by login or refresh.
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: HTTP-only refresh session cookie set by login.
  parameters:
...
    CloudflareCountry:
      name: CF-IPCountry
      in: header
      description: Cloudflare two-letter country code. IN selects INR display prices; every other value selects USD.
      schema: { type: string, minLength: 2, maxLength: 2, example: IN }
    CountryCode:
      name: X-Country-Code
      in: header
      description: Fallback country code when CF-IPCountry is absent. IN selects INR display prices; every other value selects USD.
      schema: { type: string, minLength: 2, maxLength: 2, example: IN }
  responses:
...
    BadRequest:
      description: Invalid input
      content: &errorContent
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
        text/plain:
          schema: { $ref: "#/components/schemas/PlainTextError" }
    Unauthorized:
      description: Missing or invalid authentication
      content: *errorContent
...
    InternalError:
      description: Unexpected server error
      content: *errorContent
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, example: invalid request body }
        details: { type: string }
    PlainTextError:
      type: string
      description: Error body written by net/http's http.Error. Some handlers currently put JSON-looking text in this plain-text body.
      example: invalid request body
...
    Money:
      type: object
      required: [amount_minor, amount_major, currency]
      properties:
        amount_minor: { type: integer, format: int64, description: Amount in the currency's smallest unit }
        amount_major: { type: number, format: double, description: Decimal amount in the currency's major unit }
        currency: { type: string, example: INR }
...
    Genre:
      type: object
      required: [id, name, slug, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        created_at: { type: string, format: date-time }
    LicenseOption:
      type: object
      required: [id, spec_id, type, name, price, price_money, display_price_money, features, file_types, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        spec_id: { type: string, format: uuid }
        type: { type: string, enum: [Basic, Premium, Trackout, Unlimited] }
        name: { type: string }
        price: { type: number, format: double }
        price_money: { $ref: "#/components/schemas/Money" }
        display_price_money: { $ref: "#/components/schemas/Money" }
        features: { type: array, items: { type: string } }
        file_types: { type: array, items: { type: string } }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    PublicAnalytics:
      type: object
      required: [play_count, favorite_count, total_download_count, is_favorited]
      properties:
        play_count: { type: integer }
        favorite_count: { type: integer }
        total_download_count: { type: integer }
        is_favorited: { type: boolean }
    Spec:
      type: object
      required: [id, producer_id, producer_name, title, category, type, bpm, key, description, image_url, preview_url, price, price_money, display_price_money, duration, free_mp3_enabled, created_at, updated_at, processing_status]
      properties:
        id: { type: string, format: uuid }
        producer_id: { type: string, format: uuid }
        producer_name: { type: string }
        title: { type: string }
        category: { type: string, enum: [beat, sample] }
        type: { type: string }
        bpm: { type: integer }
        key: { type: string }
        description: { type: string }
        image_url: { type: string, format: uri }
        preview_url: { type: string, format: uri }
        price: { type: number, format: double }
        price_money: { $ref: "#/components/schemas/Money" }
        display_price_money: { $ref: "#/components/schemas/Money" }
        duration: { type: integer, description: Seconds }
        free_mp3_enabled: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        licenses: { type: array, items: { $ref: "#/components/schemas/LicenseOption" } }
        genres: { type: array, items: { $ref: "#/components/schemas/Genre" } }
        tags: { type: array, items: { type: string } }
        moods: { type: array, items: { type: string } }
        instruments: { type: array, items: { type: string } }
        waveform_peaks: { type: array, items: { type: integer, format: int64 } }
        analytics: { $ref: "#/components/schemas/PublicAnalytics" }
        short_code: { type: string }
        slug: { type: string }
        processing_status: { type: string, enum: [pending, processing, completed, failed] }
...
    FavoritePage:
      type: object
      additionalProperties: false
      required:
        - items
        - has_more
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/FavoriteItem"

        has_more:
          type: boolean

        next_cursor:
          type: string
          description: Present when another page is available.
    FavoriteItem:
      type: object
      additionalProperties: false
      required:
        - favorited_at
        - spec
      properties:
        favorited_at:
          type: string
          format: date-time

        spec:
          $ref: "#/components/schemas/Spec"